pyproject.toml
# This file is used to configure the project.
# Read more about the various options under:
# https://packaging.python.org/en/latest/guides/writing-pyproject-toml
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html

[build-system]
requires = ["setuptools >= 61.0", "setuptools_scm"]
build-backend = "setuptools.build_meta"

###############################################################################
#                                 Main library                                #
###############################################################################

[project]
name = "curvlinops-for-pytorch"
authors = [
    { name = "Felix Dangel" },
    { name = "Runa Eschenhagen" },
    { name = "Lukas Tatzel" },
    { name = "Andres Fernandez"},
]
urls = { Repository = "https://github.com/f-dangel/curvlinops" }
description = "scipy Linear operators for curvature matrices in PyTorch"
readme = { file = "README.md", content-type = "text/markdown; charset=UTF-8; variant=GFM" }
license = { text = "MIT" }
# Add all kinds of additional classifiers as defined under
# https://pypi.python.org/pypi?%3Aaction=list_classifiers
classifiers = [
    "Development Status :: 4 - Beta",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dynamic = ["version"]
# Dependencies of the project:
dependencies = [
    "backpack-for-pytorch>=1.6.0",
    "torch>=2.0", # introduces torch.func
    "scipy>=1.0.0", # introduces LSMR
    "numpy>=1.21.0", # compatible with modern Python versions
    "tqdm>=4.0.0", # compatible with modern Python versions
    "einops>=0.3.0", # introduces einsum
    "einconv>=0.1.0", # first release
    "linear_operator>=0.2.0", # matvec-based CG implementation in pure PyTorch
    # NOTE PyTorch uses opt-einsum, if installed, to accelerate `torch.einsum`. This is
    # crucial for many operations in curvlinops, such as eigenvalue correction in EKFAC.
    # Hence, it is an explicit dependency and removing it will deteriorate performance.
    "opt-einsum>=3.4.0",
]
# Require a specific Python version, e.g. Python 2.7 or >= 3.4
requires-python = ">=3.10"

###############################################################################
#                           Development dependencies                          #
###############################################################################

[project.optional-dependencies]
# Dependencies needed to run the tests.
test = [
    "matplotlib",
    "tueplots",
    "coveralls",
    "pytest",
    "pytest-cov",
]

# Dependencies needed for linting.
lint = [
    "ruff",
]

# Dependencies needed to build/view the documentation.
docs = [
    "memory_profiler",
    "transformers",
    "tiktoken",
    "datasets",
    "matplotlib",
    "sphinx-gallery",
    "sphinx-rtd-theme",
    "tueplots"
]

###############################################################################
#                       Development tool configurations                       #
###############################################################################
[tool.setuptools_scm]

[tool.ruff]
line-length = 88
preview = true
exclude = [
    ".eggs",
    ".git",
    ".pytest_cache",
    "docs/rtd",
    "build",
    "dist",
]

[tool.ruff.lint]
# Enable all rules from flake8 (E, F), plus additional ones including isort (I)
# and docstrings (D)
select = ["E", "F", "B", "C", "W", "B9", "PLE", "PLW", "PLR", "I", "D"]
extend-select = [
    # Documented parameter is not in the function's signature
    "DOC102",
    # return is not documented in docstring
    "DOC201",
    # Docstring should not have a returns section because the function doesn't return anything
    "DOC202",
    # yield is not documented in docstring
    "DOC402",
    # Docstring has a "Yields" section but the function doesn't yield anything
    "DOC403",
    # Raised exception missing from docstring
    "DOC501",
    # Raised exception is not explicitly raised
    "DOC502",
]
ignore = [
    # Multi-line docstring summary should start on the second line, conflicts D212
    "D213",
    # Blank line before a class docstring, conflicts D211
    "D203",
    # E501 max-line-length (replaced by B950 (max-line-length + 10%))
    "E501",
    # C408 use {} instead of dict() (ignored because pytorch uses dict)
    "C408",
    # E203 whitespace before :
    "E203",
    # E231 missing whitespace after ','
    "E231",
    # W291 trailing whitespace
    "W291",
    # E203 line break before binary operator (replaces W503)
    "E203",
    # Line break occurred after a binary operator (replaces W504)
    "E226",
    # B905 `zip()` without an explicit `strict=` parameter
    "B905",
    # D413 missing blank line after last section
    "D413",
    # Too many arguments in function definition (9 > 5)
    "PLR0913",
    # Magic value comparison
    "PLR2004",
    # Too many positional arguments
    "PLR0917",
    # Too many local variables
    "PLR0914",
    # Method does not use self
    "PLR6301",
    # Unspecified encoding in open()
    "PLW1514",
    # Unnecessary lambda
    "PLW0108",
    # Non-augmented assignment
    "PLR6104",
    # Literal membership
    "PLR6201",
    # Bad dunder method name
    "PLW3201",
    # Loop variable overwritten by assignment target
    "PLW2901",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint.per-file-ignores]
"docs/examples/basic_usage/example_*.py" = [
    # First line should end with a period
    "D400",
    # 1 blank line required between summary line and description
    "D205",
    # First line should end with a period, question mark, or exclamation point
    "D415"
]
